+++ plugins/native_dio_adapter/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+*None.*
+
+## 1.5.1
+
 - Support request cancellation for native HTTP clients via use of `AbortableRequest` (introduced in http package from version 1.5.0)
 - Add timeout handling for `sendTimeout`, `connectTimeout`, and `receiveTimeout` in `ConversionLayerAdapter`